Move manifest under asset_management
