+++ emergence_lib/src/asset_management/manifest.rs
